Add methods to pre-validate tokens for registration and reset
These are commonly handled with a controller that serves a form for
the user to continue (unlike confirm change email, activate account,
etc which are handled as an atomic operation on clickthrough). The
validateToken method allows the controller to check the request will
be valid without running the full interactor logic, so that an error
can be shown immediately rather than only when the user submits their
new password.