Fix symfony validator dependency.
7.4 hasn't been released so composer was never installing 7 series.

Validator 7 dropped the "loose" mode of email validation

Reuse previous "loose" regex patten for email validation in login request

NB, this doesn't need to be validate an email address properly as it's not a concern of the login request. Email addresses should be fully validated by registration / change email. This is mearly to prevent values that don't even look like an email resulting in unnecessary database lookups